requirements.txt
transformers>=4.6.0,<5.0.0
tokenizers>=0.10.3
tqdm
torch>=1.6.0
torchvision
numpy~=1.25.2
scikit-learn
scipy
nltk
sentencepiece
huggingface-hub
streamlit~=1.25.0
sympy
requests~=2.31.0
openpyxl
pandas~=2.0.3
faiss-cpu
python-dotenv~=1.0.0
matplotlib~=3.7.2
requests_ntlm
Office365-REST-Python-Client
openai